+++ tests/runners/default-nil-types.sh
@@ -2,6 +2,6 @@
 
 TEST_ARGV="$1" docker-compose -f tests/docker/compose/centralized.yml \
   -f tests/docker/compose/cluster.yml \
-  -f tests/docker/runners/compose/loose-nils.yml run \
+  -f tests/docker/runners/compose/default-nil-types.yml run \
   -u $(id -u ${USER}):$(id -g ${USER}) --rm \
-  loose-nils-tests
+  default-nil-types-tests